[OPENSTACK-2441] listener support keepalive_timeout

+++ f5_openstack_agent/lbaasv2/drivers/bigip/tcp_profile.py
@@ -31,6 +31,8 @@ class TCPProfileHelper(object):
         )
         self.delete_profile = False
         self.allowed_protocols = ["TCP", "FTP", "HTTPS"]
+        self.keepalive_allowed_protocols = \
+            ["TCP", "HTTP", "HTTPS", "TERMINATED_HTTPS"]
 
     def enable_tcp(self, service):
         # pzhang: do not check ipProtocol TCP for further requirements,
@@ -284,18 +286,24 @@ class TCPProfileHelper(object):
             service, side)
         profile = "/" + partition + "/" + profile_name
 
-        LOG.info(
-            "Remove customized TCP profile: {} from "
-            "BIGIP: {}".format(
-                profile, bigip.hostname
-            )
-        )
-
-        self.tcp_helper.delete(
+        profile_exists = self.tcp_helper.exists(
             bigip,
             name=profile_name,
             partition=partition
         )
+        if profile_exists:
+            LOG.info(
+                "Remove customized TCP profile: {} from "
+                "BIGIP: {}".format(
+                    profile, bigip.hostname
+                )
+            )
+
+            self.tcp_helper.delete(
+                bigip,
+                name=profile_name,
+                partition=partition
+            )
 
     @staticmethod
     def get_profile_name(service, side):
@@ -329,3 +337,113 @@ class TCPProfileHelper(object):
         for pf in args:
             new_profiles.append(pf)
         return new_profiles
+
+    def enable_keepalive(self, service):
+        listener = service.get('listener')
+        if listener is None:
+            return False
+
+        if listener.get('keepalive_timeout') != -1:
+            protocol = listener.get('protocol')
+            if protocol not in self.keepalive_allowed_protocols:
+                return False
+            return True
+        else:
+            return False
+
+    def need_update_keepalive(self, old_listener, listener):
+        if old_listener is None or listener is None:
+            return False
+        protocol = listener.get('protocol')
+        if protocol not in self.keepalive_allowed_protocols:
+            return False
+
+        old_t = old_listener['keepalive_timeout']
+        new_t = listener['keepalive_timeout']
+        LOG.debug("keepalive_timeout old_t is {}, new_t is {}"
+                  .format(old_t, new_t))
+        if old_t == -1 and new_t == -1:
+            return False
+        else:
+            return True
+
+    def get_keepalive_timeout(self, old_listener, listener):
+        old_t = old_listener['keepalive_timeout']
+        new_t = listener['keepalive_timeout']
+        if new_t == -1:
+            return old_t
+        else:
+            return new_t
+
+    def need_delete_keepalive(self, service):
+        listener = service.get('listener')
+        if listener is None:
+            return False
+        k_t = listener.get('keepalive_timeout')
+        return True if k_t != -1 else False
+
+    def delete_keepalive_profile(self, service, vip, bigip):
+        listener = service.get('listener')
+        k_t = listener.get('keepalive_timeout')
+        if listener is not None and k_t != -1:
+            self.remove_profile(service, vip, bigip, side="server")
+
+    def add_keepalive_tcp_profile(self, service, vip, bigip, **kwargs):
+        keepalive_timeout = kwargs.get('keepalive_timeout')
+        side = kwargs.get("side")
+        tcp_options = kwargs.get("tcp_options")
+        tcp_options = "{%s first}" % tcp_options if tcp_options else None
+
+        partition = vip['partition']
+        profile_name = self.get_profile_name(service, side)
+        profile = "/" + partition + "/" + profile_name
+
+        profile_exists = self.tcp_helper.exists(
+            bigip,
+            name=profile_name,
+            partition=partition
+        )
+
+        if not profile_exists:
+            payload = dict(
+                name=profile_name,
+                partition=partition,
+                tcpOptions=tcp_options,
+                idleTimeout=keepalive_timeout,
+            )
+            LOG.info(
+                "Add customized keepalive TCP profile: {} for "
+                "BIGIP: {} ".format(
+                    profile, bigip.hostname
+                )
+            )
+            self.tcp_helper.create(bigip, payload)
+        else:
+            profile = self.tcp_helper.load(bigip, name=profile_name,
+                                           partition=partition)
+            profile.modify(idleTimeout=keepalive_timeout)
+
+        server_profile_body = {
+            "name": profile_name,
+            "partition": partition,
+            "context": "serverside"
+        }
+        client_profile_body = {
+            "name": "tcp",
+            "partition": "Common",
+            "context": "clientside"
+        }
+        if vip.get('profiles'):
+            delete_fastL4s = vip['profiles'].count("/Common/fastL4")
+            for _ in range(delete_fastL4s):
+                vip['profiles'].remove("/Common/fastL4")
+
+            if server_profile_body not in vip['profiles']:
+                vip['profiles'].append(server_profile_body)
+            if client_profile_body not in vip['profiles']:
+                vip['profiles'].append(client_profile_body)
+        else:
+            profiles = [server_profile_body, client_profile_body]
+            vip['profiles'] = profiles
+
+        self.delete_profile = False
